ECNG-1560: fix

+++ EasyCard/Transactions.Api/Models/Masav/MasavFileRowSummary.cs
@@ -16,11 +16,11 @@ namespace Transactions.Api.Models.Masav
 
         public Guid? PaymentTransactionID { get; set; }
 
-        public int? Bankcode { get; set; }
+        public long? Bankcode { get; set; }
 
-        public int? BranchNumber { get; set; }
+        public long? BranchNumber { get; set; }
 
-        public int? AccountNumber { get; set; }
+        public long? AccountNumber { get; set; }
 
         [MetadataOptions(Hidden = true)]
         public string NationalID { get; set; }
